+++ src/Services/BitcoinService.cs
@@ -193,6 +193,9 @@ namespace FundsManager.Services
                 }
 
                 result.Item1 = builder.BuildPSBT(false);
+                
+                //Additional fields to support PSBT signing with a HW or the Remote Signer 
+                LightningHelper.SetGlobalXPUBMetadata(_logger,walletWithdrawalRequest.Wallet.Keys, result, selectedUTXOs, scriptCoins);
             }
             catch (Exception e)
             {
